loan_application_agent/requirements.txt
google-adk>=1.0.0
litellm>=1.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-cloud-aiplatform[adk]>=1.76.0
pydantic-settings>=2.7.0
